Version 2.0.0-dev
- Handling each component and collector in separate files
- Optimize code
