+++ src/libs/collector/pageNumber.js
@@ -1,0 +1,3 @@
+let page = 0;
+
+module.exports.page;
